Move diff up one level so it's accessible by other packages
