+++ internal/diff/adapter_test.go
@@ -1,4 +1,4 @@
-package kube
+package diff
 
 import (
 	"testing"
