+++ internal/diff/diff.go
@@ -1,4 +1,4 @@
-package kube
+package diff
 
 import (
 	"errors"
