+++ internal/diff/diff_test.go
@@ -1,4 +1,4 @@
-package kube
+package diff
 
 import (
 	"fmt"
